+++ reikna/linalg/__init__.py
@@ -1,0 +1,20 @@
+"""
+Linear algebra algorithms.
+
+
+Matrix multiplication (dot product)
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. autoclass:: MatrixMul
+    :members:
+
+
+Matrix norms
+^^^^^^^^^^^^
+
+.. autoclass:: EntrywiseNorm
+    :members:
+"""
+
+from reikna.linalg.matrixmul import MatrixMul
+from reikna.linalg.norm import EntrywiseNorm