+++ reikna/transformations.py
@@ -25,37 +25,65 @@ def copy(arr_t, out_arr_t=None):
         "${output.store_same}(${input.load_same});")
 
 
-def scale_param(arr_t, coeff_dtype):
+def add_param(arr_t, param_dtype):
     """
-    Returns a scaling transformation with dynamic parameter (1 output, 1 input, 1 scalar):
-    ``output = input * coeff``.
+    Returns an addition transformation with a dynamic parameter (1 output, 1 input, 1 scalar):
+    ``output = input + param``.
     """
     return Transformation(
         [Parameter('output', Annotation(arr_t, 'o')),
         Parameter('input', Annotation(arr_t, 'i')),
-        Parameter('coeff', Annotation(coeff_dtype))],
-        "${output.store_same}(${mul}(${input.load_same}, ${coeff}));",
-        render_kwds=dict(mul=functions.mul(arr_t.dtype, coeff_dtype, out_dtype=arr_t.dtype)))
+        Parameter('param', Annotation(param_dtype))],
+        "${output.store_same}(${add}(${input.load_same}, ${param}));",
+        render_kwds=dict(add=functions.add(arr_t.dtype, param_dtype, out_dtype=arr_t.dtype)))
 
 
-def scale_const(arr_t, coeff):
+def add_const(arr_t, param):
     """
-    Returns a scaling transformation with fixed parameter (1 output, 1 input):
-    ``output = input * <coeff>``.
+    Returns an addition transformation with a fixed parameter (1 output, 1 input):
+    ``output = input + param``.
     """
-    coeff_dtype = dtypes.detect_type(coeff)
+    param_dtype = dtypes.detect_type(param)
     return Transformation(
         [Parameter('output', Annotation(arr_t, 'o')),
         Parameter('input', Annotation(arr_t, 'i'))],
-        "${output.store_same}(${mul}(${input.load_same}, ${coeff}));",
+        "${output.store_same}(${add}(${input.load_same}, ${param}));",
         render_kwds=dict(
-            mul=functions.mul(arr_t.dtype, coeff_dtype, out_dtype=arr_t.dtype),
-            coeff=dtypes.c_constant(coeff, dtype=coeff_dtype)))
+            add=functions.add(arr_t.dtype, param_dtype, out_dtype=arr_t.dtype),
+            param=dtypes.c_constant(param, dtype=param_dtype)))
+
+
+def mul_param(arr_t, param_dtype):
+    """
+    Returns a scaling transformation with a dynamic parameter (1 output, 1 input, 1 scalar):
+    ``output = input * param``.
+    """
+    return Transformation(
+        [Parameter('output', Annotation(arr_t, 'o')),
+        Parameter('input', Annotation(arr_t, 'i')),
+        Parameter('param', Annotation(param_dtype))],
+        "${output.store_same}(${mul}(${input.load_same}, ${param}));",
+        render_kwds=dict(mul=functions.mul(arr_t.dtype, param_dtype, out_dtype=arr_t.dtype)))
+
+
+def mul_const(arr_t, param):
+    """
+    Returns a scaling transformation with a fixed parameter (1 output, 1 input):
+    ``output = input * param``.
+    """
+    param_dtype = dtypes.detect_type(param)
+    return Transformation(
+        [Parameter('output', Annotation(arr_t, 'o')),
+        Parameter('input', Annotation(arr_t, 'i'))],
+        "${output.store_same}(${mul}(${input.load_same}, ${param}));",
+        render_kwds=dict(
+            mul=functions.mul(arr_t.dtype, param_dtype, out_dtype=arr_t.dtype),
+            param=dtypes.c_constant(param, dtype=param_dtype)))
 
 
 def split_complex(input_arr_t):
     """
-    Returns a transformation which splits complex input into two real outputs
+    Returns a transformation that splits complex input into two real outputs
     (2 outputs, 1 input): ``real = Re(input), imag = Im(input)``.
     """
     output_t = Type(dtypes.real_for(input_arr_t.dtype), shape=input_arr_t.shape)
@@ -71,7 +99,7 @@ def split_complex(input_arr_t):
 
 def combine_complex(output_arr_t):
     """
-    Returns a transformation which joins two real inputs into complex output
+    Returns a transformation that joins two real inputs into complex output
     (1 output, 2 inputs): ``output = real + 1j * imag``.
     """
     input_t = Type(dtypes.real_for(output_arr_t.dtype), shape=output_arr_t.shape)
@@ -84,4 +112,99 @@ def combine_complex(output_arr_t):
             COMPLEX_CTR(${output.ctype})(
                 ${real.load_same},
                 ${imag.load_same}));
+        """)
+
+
+def norm_const(arr_t, order):
+    """
+    Returns a transformation that calculates the ``order``-norm
+    (1 output, 1 input): ``output = abs(input) ** order``.
+    """
+    if dtypes.is_complex(arr_t.dtype):
+        out_dtype = dtypes.real_for(arr_t.dtype)
+    else:
+        out_dtype = arr_t.dtype
+
+    return Transformation(
+        [
+            Parameter('output', Annotation(Type(out_dtype, arr_t.shape), 'o')),
+            Parameter('input', Annotation(arr_t, 'i'))],
+        """
+        ${input.ctype} val = ${input.load_same};
+        ${output.ctype} norm = ${norm}(val);
+        %if order != 2:
+        norm = pow(norm, ${dtypes.c_constant(order / 2, output.dtype)});
+        %endif
+        ${output.store_same}(norm);
+        """,
+        render_kwds=dict(
+            norm=functions.norm(arr_t.dtype),
+            order=order))
+
+
+def norm_param(arr_t):
+    """
+    Returns a transformation that calculates the ``order``-norm
+    (1 output, 1 input, 1 param): ``output = abs(input) ** order``.
+    """
+    if dtypes.is_complex(arr_t.dtype):
+        out_dtype = dtypes.real_for(arr_t.dtype)
+    else:
+        out_dtype = arr_t.dtype
+
+    return Transformation(
+        [
+            Parameter('output', Annotation(Type(out_dtype, arr_t.shape), 'o')),
+            Parameter('input', Annotation(arr_t, 'i')),
+            Parameter('order', Annotation(Type(out_dtype)))],
+        """
+        ${input.ctype} val = ${input.load_same};
+        ${output.ctype} norm = ${norm}(val);
+        norm = pow(norm, ${order} / 2);
+        ${output.store_same}(norm);
+        """,
+        render_kwds=dict(
+            norm=functions.norm(arr_t.dtype)))
+
+
+def ignore(arr_t):
+    """
+    Returns a transformation that ignores the output it is attached to.
+    """
+    return Transformation(
+        [Parameter('input', Annotation(arr_t, 'i'))],
+        """
+        // Ignoring intentionally
+        """)
+
+
+def broadcast_const(arr_t, val):
+    """
+    Returns a transformation that broadcasts the given constant to the array output
+    (1 output): ``output = val``.
+    """
+    val = dtypes.cast(arr_t.dtype)(val)
+    if len(val.shape) != 0:
+        raise ValueError("The constant must be a scalar")
+    return Transformation(
+        [
+            Parameter('output', Annotation(arr_t, 'o'))],
+        """
+        const ${output.ctype} val = ${dtypes.c_constant(val)};
+        ${output.store_same}(val);
+        """,
+        render_kwds=dict(val=val))
+
+
+def broadcast_param(arr_t):
+    """
+    Returns a transformation that broadcasts the free parameter to the array output
+    (1 output, 1 param): ``output = param``.
+    """
+    return Transformation(
+        [
+            Parameter('output', Annotation(arr_t, 'o')),
+            Parameter('param', Annotation(Type(arr_t.dtype)))],
+        """
+        ${output.store_same}(${param});
         """)